refactor: reorganize tasks by category package structure for improved clarity
