+++ app/src/main/java/com/giraffe/tudeeapp/MainActivity.kt
@@ -18,16 +18,14 @@ import androidx.navigation.navArgument
 import com.giraffe.tudeeapp.design_system.component.NavBar
 import com.giraffe.tudeeapp.design_system.theme.TudeeTheme
 import com.giraffe.tudeeapp.presentation.categories.screens.CategoriesScreen
-import com.giraffe.tudeeapp.presentation.categories.tasks_by_category.TasksByCategoryScreen
-import com.giraffe.tudeeapp.presentation.categories.tasks_by_category.TasksByCategoryViewModel
+import com.giraffe.tudeeapp.presentation.tasks_by_category.TasksByCategoryScreen
+import com.giraffe.tudeeapp.presentation.tasks_by_category.TasksByCategoryViewModel
 import com.giraffe.tudeeapp.presentation.home.HomeScreen
 import com.giraffe.tudeeapp.presentation.navigation.Screen
 import com.giraffe.tudeeapp.presentation.splash.onboard.OnboardingScreen
 import com.giraffe.tudeeapp.presentation.splash.splashscreen.SplashScreen
 import com.giraffe.tudeeapp.presentation.tasks.TaskScreen
-import org.koin.android.ext.android.getKoin
 import org.koin.androidx.compose.koinViewModel
-import org.koin.core.parameter.parametersOf
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
